+++ memori/utils/logging.py
@@ -26,11 +26,13 @@ class LoggingManager:
             if not cls._initialized:
                 logger.remove()
 
-            if verbose:
-                # When verbose mode is enabled, disable all other loggers and show only loguru
-                cls._disable_other_loggers()
+            # CRITICAL: Always disable third-party loggers to prevent performance degradation
+            # SQLAlchemy, OpenAI, and other libraries have verbose logging by default
+            # which can cause 30-70% performance overhead if not disabled
+            cls._disable_other_loggers()
 
-                # Configure console logging with DEBUG level and full formatting
+            if verbose:
+                # When verbose mode is enabled, show Memori debug logging
                 logger.add(
                     sys.stderr,
                     level="DEBUG",